% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts_bytecode.R
\name{contracts_bytecode}
\alias{contracts_bytecode}
\title{Retrieve smart contract bytecode}
\usage{
contracts_bytecode(config, contract_id, .transport = NULL)
}
\arguments{
\item{config}{A configuration list from \code{\link[=hadeda_config]{hadeda_config()}}.}

\item{contract_id}{Contract identifier.}

\item{.transport}{Optional transport override.}
}
\value{
A tibble with the contract identifier and bytecode.
}
\description{
Fetch bytecode for a smart contract from the Mirror Node.
}
\examples{
config <- hadeda_config()
contract_id <- "0.0.5005"
bytecode <- contracts_bytecode(config, contract_id)
bytecode

}
