% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts_get.R
\name{contracts_get}
\alias{contracts_get}
\title{Retrieve contract information}
\usage{
contracts_get(config, contract_id, .transport = NULL)
}
\arguments{
\item{config}{A configuration list from \code{\link[=hadeda_config]{hadeda_config()}}.}

\item{contract_id}{Contract identifier.}

\item{.transport}{Optional transport override.}
}
\value{
A tibble describing the contract.
}
\description{
Fetch smart contract metadata from the Mirror Node.
}
\examples{
config <- hadeda_config()
contract_id <- "0.0.5005"
contract <- contracts_get(config, contract_id)
contract

}
