% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_delete}
\alias{file_delete}
\title{Delete a Hedera file}
\usage{
file_delete(
  config,
  file_id,
  max_fee = NULL,
  wait_for_receipt = TRUE,
  .transport = NULL
)
}
\arguments{
\item{config}{A configuration list created by \code{\link[=hadeda_config]{hadeda_config()}}.}

\item{file_id}{File identifier to update.}

\item{max_fee}{Optional fee ceiling expressed in tinybars.}

\item{wait_for_receipt}{Logical indicating whether the gRPC handler should
wait for a transaction receipt acknowledgement.}

\item{.transport}{Optional transport override. Only the gRPC transport is
supported.}
}
\description{
Submit a FileService \code{deleteFile} transaction to mark a file as deleted.
}
