% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_coalesce.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing helper}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{Values to compare.}
}
\value{
\code{x} when it is available, otherwise \code{y}.
}
\description{
Return the left-hand side if it is neither \code{NULL} nor length zero, otherwise return the right-hand side.
}
\examples{
"value" \%||\% "fallback"
NULL \%||\% "fallback"

}
\keyword{internal}
