% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda-package.R
\docType{package}
\name{hadeda}
\alias{hadeda-package}
\alias{hadeda}
\title{Hadeda: Hedera Hashgraph Tidyverse Client}
\description{
The Hadeda package provides tidyverse-friendly wrappers around the
Hedera Hashgraph Mirror Node REST API and gRPC services. The
exported helpers orchestrate authentication, transport selection,
and response normalisation so that analysts can compose pipelines
with familiar tidyverse verbs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hadeda-r/hadeda}
  \item Report bugs at \url{https://github.com/hadeda-r/hadeda/issues}
}

}
\author{
\strong{Maintainer}: Christiaan Pauw \email{cjpauw@gmail.com}

}
\keyword{internal}
