% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda_network_defaults.R
\name{hadeda_network_defaults}
\alias{hadeda_network_defaults}
\title{Default configuration for known Hedera networks}
\usage{
hadeda_network_defaults(network)
}
\arguments{
\item{network}{One of "mainnet", "testnet", "previewnet", or "localhost".}
}
\value{
A list containing default \code{rest} and \code{grpc} configurations.
}
\description{
Provide sensible transport defaults for named Hedera networks.
}
\examples{
hadeda_network_defaults("testnet")

}
\keyword{internal}
