% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda_parse_account_allowances_nfts.R
\name{hadeda_parse_account_allowances_nfts}
\alias{hadeda_parse_account_allowances_nfts}
\title{Parse NFT allowance records}
\usage{
hadeda_parse_account_allowances_nfts(records)
}
\arguments{
\item{records}{A list of NFT allowance entries from the Mirror Node REST API.}
}
\value{
A tibble describing NFT allowance approvals.
}
\description{
Convert NFT allowance payloads from the Mirror Node into a tidy tibble.
}
\examples{
records <- list(list(
  owner = "0.0.1001",
  token_id = "0.0.2002",
  granted = list(list(
    spender = "0.0.3003",
    serial_numbers = list(1, 2),
    timestamp = "1672531200.000000000",
    payer_account_id = "0.0.4004",
    transaction_id = "0.0.4004-123456789-000000000"
  )),
  approved_for_all = list(list(
    spender = "0.0.5005",
    timestamp = "1672531300.000000000",
    delegating_spender = "0.0.6006",
    payer_account_id = "0.0.7007",
    transaction_id = "0.0.7007-123456789-000000000"
  ))
))
hadeda_parse_account_allowances_nfts(records)

}
\keyword{internal}
