% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda_parse_account_balances.R
\name{hadeda_parse_account_balances}
\alias{hadeda_parse_account_balances}
\title{Parse account balance responses}
\usage{
hadeda_parse_account_balances(resp)
}
\arguments{
\item{resp}{A list response from \code{\link[=hadeda_rest_get]{hadeda_rest_get()}} for the account balance endpoint.}
}
\value{
A tibble of balances by account.
}
\description{
Convert balance payloads returned from \code{accounts/{id}/balance} to a tibble.
}
\examples{
resp <- list(
  timestamp = "1700000000.000000000",
  balances = list(list(account = "0.0.1001", balance = 100L, tokens = list()))
)
hadeda_parse_account_balances(resp)

}
\keyword{internal}
