% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda_parse_blocks.R
\name{hadeda_parse_blocks}
\alias{hadeda_parse_blocks}
\title{Parse block summaries}
\usage{
hadeda_parse_blocks(records)
}
\arguments{
\item{records}{A list of block objects as returned by the Mirror Node REST API.}
}
\value{
A tibble describing blocks.
}
\description{
Transform raw block records from the Mirror Node API into a tibble.
}
\examples{
records <- list(list(
  number = 1000L,
  hash = "0xabc",
  previous_hash = "0xdef",
  count = 25,
  start_consensus_timestamp = "1700000000.000000000",
  end_consensus_timestamp = "1700000001.000000000"
))
hadeda_parse_blocks(records)

}
\keyword{internal}
