% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda_parse_network_stake.R
\name{hadeda_parse_network_stake}
\alias{hadeda_parse_network_stake}
\title{Parse network stake snapshots}
\usage{
hadeda_parse_network_stake(records)
}
\arguments{
\item{records}{A list of stake snapshots.}
}
\value{
A tibble summarising network stake.
}
\description{
Convert network stake payloads from the Mirror Node API into a tibble.
}
\keyword{internal}
