% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda_parse_schedules.R
\name{hadeda_parse_schedule_signatures}
\alias{hadeda_parse_schedule_signatures}
\title{Parse schedule signatures}
\usage{
hadeda_parse_schedule_signatures(signatures)
}
\arguments{
\item{signatures}{A list of schedule signature entries.}
}
\value{
A tibble of signatures.
}
\description{
Convert schedule signature payloads into a tibble.
}
\keyword{internal}
