% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda_parse_timestamp.R
\name{hadeda_parse_timestamp}
\alias{hadeda_parse_timestamp}
\title{Parse consensus timestamps}
\usage{
hadeda_parse_timestamp(x)
}
\arguments{
\item{x}{A character vector of consensus timestamps.}
}
\value{
A POSIXct vector in UTC.
}
\description{
Convert Hedera consensus timestamps to POSIXct values.
}
\examples{
hadeda_parse_timestamp(c("1700000000.000000001", "1700000001.5"))

}
\keyword{internal}
