% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda_parse_token_balances.R
\name{hadeda_parse_token_balances}
\alias{hadeda_parse_token_balances}
\title{Parse token balance records}
\usage{
hadeda_parse_token_balances(records)
}
\arguments{
\item{records}{A list of token balance entries.}
}
\value{
A tibble of token balances.
}
\description{
Convert token balance payloads into a tidy tibble.
}
\examples{
records <- list(list(account = "0.0.1001", balance = 100L, decimals = 8))
hadeda_parse_token_balances(records)

}
\keyword{internal}
