% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda_parse_token_nfts.R
\name{hadeda_parse_token_nfts}
\alias{hadeda_parse_token_nfts}
\title{Parse token NFT records}
\usage{
hadeda_parse_token_nfts(records)
}
\arguments{
\item{records}{A list of NFT objects as returned by the Mirror Node REST API.}
}
\value{
A tibble describing token NFTs.
}
\description{
Transform raw NFT records from the Mirror Node API into a tidy tibble.
}
\examples{
records <- list(list(
  token_id = "0.0.1001",
  serial_number = 1,
  account_id = "0.0.2002",
  created_timestamp = "1700000000.000000000",
  metadata = "aGVsbG8="
))
hadeda_parse_token_nfts(records)

}
\keyword{internal}
