% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda_parse_tokens.R
\name{hadeda_parse_tokens}
\alias{hadeda_parse_tokens}
\title{Parse token metadata}
\usage{
hadeda_parse_tokens(records)
}
\arguments{
\item{records}{A list of token metadata objects.}
}
\value{
A tibble describing tokens.
}
\description{
Transform raw token records from the Mirror Node into a tibble.
}
\examples{
records <- list(list(
  token_id = "0.0.1001",
  symbol = "HED",
  name = "Example Token",
  treasury_account_id = "0.0.3",
  type = "FUNGIBLE_COMMON"
))
hadeda_parse_tokens(records)

}
\keyword{internal}
