% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda_parse_topic_messages.R
\name{hadeda_parse_topic_messages}
\alias{hadeda_parse_topic_messages}
\title{Parse topic message records}
\usage{
hadeda_parse_topic_messages(records)
}
\arguments{
\item{records}{A list of topic message records.}
}
\value{
A tibble describing topic messages.
}
\description{
Convert Mirror Node topic messages into a tibble.
}
\examples{
records <- list(list(
  topic_id = "0.0.2002",
  consensus_timestamp = "1700000000.000000001",
  message = "SGVsbG8=",
  running_hash = "abcd",
  sequence_number = 1
))
hadeda_parse_topic_messages(records)

}
\keyword{internal}
