% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda_parse_transactions.R
\name{hadeda_parse_transactions}
\alias{hadeda_parse_transactions}
\title{Parse transaction records}
\usage{
hadeda_parse_transactions(records)
}
\arguments{
\item{records}{A list of transaction records.}
}
\value{
A tibble describing transactions.
}
\description{
Convert Mirror Node transaction payloads into a tibble.
}
\examples{
records <- list(list(
  transaction_id = "0.0.1001-1700000000-000000000",
  consensus_timestamp = "1700000000.000000001",
  name = "CRYPTOTRANSFER",
  valid_start_timestamp = "1700000000.000000000",
  charged_tx_fee = 1000
))
hadeda_parse_transactions(records)

}
\keyword{internal}
