% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda_proto_bundle.R
\name{hadeda_proto_bundle_version}
\alias{hadeda_proto_bundle_version}
\title{Report the version of the local Hedera protobuf bundle, if any}
\usage{
hadeda_proto_bundle_version()
}
\value{
A single string containing the downloaded version, or \code{NA} when no
bundle is available.
}
\description{
Report the version of the local Hedera protobuf bundle, if any
}
