% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadeda_require_rest.R
\name{hadeda_require_rest}
\alias{hadeda_require_rest}
\title{Ensure a REST configuration is present}
\usage{
hadeda_require_rest(config, caller = rlang::caller_env())
}
\arguments{
\item{config}{A configuration list from \code{\link[=hadeda_config]{hadeda_config()}}.}

\item{caller}{Environment used for error messaging.}
}
\value{
The REST configuration list.
}
\description{
Validate that a configuration includes REST settings.
}
\examples{
config <- hadeda_config()
hadeda_require_rest(config)

}
\keyword{internal}
