% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_exchange_rate.R
\name{network_exchange_rate}
\alias{network_exchange_rate}
\title{Retrieve Hedera network exchange rates}
\usage{
network_exchange_rate(config, .transport = NULL)
}
\arguments{
\item{config}{A configuration list created by \code{\link[=hadeda_config]{hadeda_config()}}.}

\item{.transport}{Optional transport override ("rest" or "grpc").}
}
\value{
A tibble describing the current and next exchange rates.
}
\description{
Fetch the current and next HBAR to USD cent exchange rates from the Mirror Node REST API.
}
\examples{
mirror <- hadeda_config(network = "testnet")
\dontrun{
  network_exchange_rate(mirror)
}

}
