% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedules.R
\name{schedules_list}
\alias{schedules_list}
\title{List scheduled transactions}
\usage{
schedules_list(
  config,
  limit = NULL,
  order = c("desc", "asc"),
  schedule_id = NULL,
  transaction_id = NULL,
  scheduled_timestamp = NULL,
  executed_timestamp = NULL,
  deleted = NULL,
  timestamp = NULL,
  .transport = NULL
)
}
\arguments{
\item{config}{A configuration list from \code{\link[=hadeda_config]{hadeda_config()}}.}

\item{limit}{Optional page size.}

\item{order}{Ordering of results, either "desc" or "asc".}

\item{schedule_id}{Optional schedule identifier filter.}

\item{transaction_id}{Optional scheduling transaction identifier filter.}

\item{scheduled_timestamp}{Optional scheduled execution timestamp filter.}

\item{executed_timestamp}{Optional executed timestamp filter.}

\item{deleted}{Optional logical flag to restrict deleted vs active schedules.}

\item{timestamp}{Optional timestamp range filter.}

\item{.transport}{Optional transport override.}
}
\value{
A tibble describing scheduled transactions.
}
\description{
Retrieve scheduled transactions from the Hedera Mirror Node REST API with
optional filters.
}
