% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_balances.R
\name{tokens_balances}
\alias{tokens_balances}
\title{Retrieve token balance information}
\usage{
tokens_balances(
  config,
  token_id,
  limit = NULL,
  account_id = NULL,
  order = c("desc", "asc"),
  .transport = NULL
)
}
\arguments{
\item{config}{A configuration list from \code{\link[=hadeda_config]{hadeda_config()}}.}

\item{token_id}{Token identifier.}

\item{limit}{Optional page size.}

\item{account_id}{Optional filter restricting balances to a specific account.}

\item{order}{Ordering of results, either "desc" or "asc".}

\item{.transport}{Optional transport override.}
}
\value{
A tibble of token balances.
}
\description{
Fetch account balances for a token from the Mirror Node.
}
\examples{
mirror <- hadeda_config(network = "testnet")
hashio <- hadeda_config(
  network = "testnet",
  rest = list(
    base_url = "https://testnet.hashio.io/api/v1",
    headers = list(`X-API-Key` = Sys.getenv("HASHIO_API_KEY"))
  ),
  default_transport = "rest"
)
\dontrun{
  new_account <- accounts_create(hashio)
  new_token <- tokens_create(
    hashio,
    name = "Hadeda Example",
    symbol = "HADEDA",
    treasury_account_id = new_account$account
  )
  tokens_balances(
    config = mirror,
    token_id = new_token$token_id,
    limit = 10,
    account_id = new_account$account
  )
}

}
