% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_get.R
\name{tokens_get}
\alias{tokens_get}
\title{Retrieve token metadata}
\usage{
tokens_get(config, token_id, .transport = NULL)
}
\arguments{
\item{config}{A configuration list from \code{\link[=hadeda_config]{hadeda_config()}}.}

\item{token_id}{Token identifier.}

\item{.transport}{Optional transport override.}
}
\value{
A tibble describing the token.
}
\description{
Fetch token details from the Mirror Node.
}
\examples{
mirror <- hadeda_config(network = "testnet")
hashio <- hadeda_config(
  network = "testnet",
  rest = list(
    base_url = "https://testnet.hashio.io/api/v1",
    headers = list(`X-API-Key` = Sys.getenv("HASHIO_API_KEY"))
  ),
  default_transport = "rest"
)
\dontrun{
  new_account <- accounts_create(hashio)
  new_token <- tokens_create(
    hashio,
    name = "Hadeda Example",
    symbol = "HADEDA",
    treasury_account_id = new_account$account
  )
  tokens_get(mirror, new_token$token_id)
}

}
