% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_nfts.R
\name{tokens_nft}
\alias{tokens_nft}
\title{Retrieve a specific token NFT}
\usage{
tokens_nft(config, token_id, serial_number, .transport = NULL)
}
\arguments{
\item{config}{A configuration list from \code{\link[=hadeda_config]{hadeda_config()}}.}

\item{token_id}{Token identifier to query.}

\item{serial_number}{Serial number of the NFT.}

\item{.transport}{Optional transport override.}
}
\value{
A tibble with a single NFT record.
}
\description{
Retrieve a specific token NFT
}
